Improved sound managment

+++ com/BombingGames/EngineCore/View.java
@@ -76,15 +76,19 @@ public class View {
         
         
         //render HUD
-       // hudCamera.zoom = 1/equalizationScale;
+        // hudCamera.zoom = 1/equalizationScale;
         hudCamera.update();
         hudCamera.apply(Gdx.gl10);
          
         batch.setProjectionMatrix(hudCamera.combined);
         shapeRenderer.setProjectionMatrix(hudCamera.combined);
         
-        Gdx.gl.glViewport((int) 0, (int) 0,
-                          (int) Gdx.graphics.getWidth(), (int) Gdx.graphics.getHeight());
+        Gdx.gl.glViewport(
+            0,
+            0,
+            (int) Gdx.graphics.getWidth(),
+            (int) Gdx.graphics.getHeight()
+        );
         
         drawString("FPS:"+ Gdx.graphics.getFramesPerSecond(), 10, 10);
         
@@ -92,7 +96,7 @@ public class View {
         //hudCamera.zoom = 1/equalizationScale;
         
         if (Controller.getLightengine() != null)
-        Controller.getLightengine().render(this);
+            Controller.getLightengine().render(this);
         
         if (controller.getMinimap() != null)
             controller.getMinimap().render(this);
